feat: SKSK-59 모킹함수들 추상화
